+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 All notable changes to this project are documented here.
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Added
+- `BootstrapDistribution.summarize` has new parameter `precision: Optional[Union[int, Literal["auto"]]] = None` to control rounding behaviour.
+
+### Changed
+- `BootstrapSummary.round` parameter `precision` now defaults to `Literal["auto"]`.
 
 ## [0.1.1] - 2025-11-03